+++ packages/bsky/src/api/index.ts
@@ -3,7 +3,7 @@ import AppContext from '../context'
 import describeFeedGenerator from './app/bsky/feed/describeFeedGenerator'
 import getTimeline from './app/bsky/feed/getTimeline'
 import getActorFeeds from './app/bsky/feed/getActorFeeds'
-import getFeedSuggestions from './app/bsky/feed/getFeedSuggestions'
+import getSuggestedFeeds from './app/bsky/feed/getSuggestedFeeds'
 import getAuthorFeed from './app/bsky/feed/getAuthorFeed'
 import getFeed from './app/bsky/feed/getFeed'
 import getFeedGenerator from './app/bsky/feed/getFeedGenerator'
@@ -61,7 +61,7 @@ export default function (server: Server, ctx: AppContext) {
   describeFeedGenerator(server, ctx)
   getTimeline(server, ctx)
   getActorFeeds(server, ctx)
-  getFeedSuggestions(server, ctx)
+  getSuggestedFeeds(server, ctx)
   getAuthorFeed(server, ctx)
   getFeed(server, ctx)
   getFeedGenerator(server, ctx)
