+++ packages/pds/src/app-view/api/app/bsky/index.ts
@@ -2,7 +2,7 @@ import { Server } from '../../../../lexicon'
 import AppContext from '../../../../context'
 import getTimeline from './feed/getTimeline'
 import getActorFeeds from './feed/getActorFeeds'
-import getFeedSuggestions from './feed/getFeedSuggestions'
+import getSuggestedFeeds from './feed/getSuggestedFeeds'
 import getAuthorFeed from './feed/getAuthorFeed'
 import getFeedGenerator from './feed/getFeedGenerator'
 import getFeedGenerators from './feed/getFeedGenerators'
@@ -38,7 +38,7 @@ import unspecced from './unspecced'
 export default function (server: Server, ctx: AppContext) {
   getTimeline(server, ctx)
   getActorFeeds(server, ctx)
-  getFeedSuggestions(server, ctx)
+  getSuggestedFeeds(server, ctx)
   getAuthorFeed(server, ctx)
   getFeedGenerator(server, ctx)
   getFeedGenerators(server, ctx)
